feat: split up code into multiple files
