+++ cli/internal/graph/renderer/render_test.go
